feat(api): extract parse_birth_date_input helper + replace both call sites [GREEN]
Add graphql/input/birth_date.rs with parse_birth_date_input() that
consolidates the 15-line BirthDateInput parsing closure. Replace the
duplicate blocks in create_dog_field and update_dog_field with a
single-line call to the helper. Add BirthDate::to_json unit tests.
